separate admin tool commands to essync
